dictionary: move iterator_wrapper and stl_dictionary_wrapper to helper module
