+++ src/dictionary/helper/iterator_wrapper.h
