+++ src/dictionary/linear_list_dictionary.h
@@ -5,7 +5,7 @@
 #include <utility>
 #include <string>
 #include <algorithm>
-#include "iterator_wrapper.h"
+#include "helper/iterator_wrapper.h"
 
 namespace dictionary {
     template<class K, class V, class...>
